prompt a lienzo

+++ src/components/prueba.jsx
@@ -17,6 +17,13 @@ export default function CreateLienzo() {
   const { roomId } = useParams();
   const navigate = useNavigate();
   const [elementsDraft, setElementsDraft] = useState([]);
+  const [htmlFinal, setHtmlFinal] = useState("");
+
+
+
+ 
+
+
 
 
   // Funciones de creación
@@ -299,9 +306,21 @@ export default function CreateLienzo() {
     .join("\n");
 
   useEffect(() => {
-    localStorage.setItem("htmlGenerado", generatedHTML);
+    const existe = localStorage.getItem("htmlGenerado");
+    if (!existe) {
+      localStorage.setItem("htmlGenerado", generatedHTML);
+    }
   }, [generatedHTML]);
 
+
+   useEffect(() => {
+    const guardado = localStorage.getItem("htmlGenerado");
+    if (guardado) {
+      setHtmlFinal(guardado);
+    } else {
+      setHtmlFinal(generatedHTML);
+    }
+  }, [generatedHTML]);
   //Sidebar
   return (
     <>
@@ -599,11 +618,12 @@ export default function CreateLienzo() {
           <h2>📄 HTML generado</h2>
           <textarea
             className="textarea-html"
-            value={generatedHTML}
+            value={htmlFinal}
             readOnly
             rows="10"
             cols="50"
           />
+
           <button onClick={() => navigate('/generar-flutter')}>
             Descargar Vista en Angular
           </button>
